requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-rest-auth==5.0.2
Django==5.0.2
django-cache-url==3.4.5
django-debug-toolbar==4.3.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
ecdsa==0.18.0
environs==10.3.0
gunicorn==21.2.0
idna==3.6
jwcrypto==1.5.4
marshmallow==3.21.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
pypaystack==1.24
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
